requirements.txt
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
openai>=1.30.0
python-dotenv>=1.0.0
